feat:代码规范优化

+++ bubble-api/bubble-api-oa/src/main/java/com/bubblecloud/oa/api/entity/EnterpriseRole.java
@@ -48,7 +48,7 @@ public class EnterpriseRole extends Req<EnterpriseRole> {
 	private String apiUnique;
 
 	@Schema(description = "企业ID")
-	private Integer entid;
+	private Long entid;
 
 	@Schema(description = "数据范围")
 	private Integer dataLevel;
